# modules/rest/openapi/openapi.yaml
openapi: 3.0.1
info:
  title: Ignite REST module
  contact:
    email: user@ignite.apache.org
  license:
    name: Apache 2.0
    url: https://ignite.apache.org
  version: 3.0.0-alpha
paths:
  /management/v1/configuration/cluster:
    get:
      tags:
      - clusterConfiguration
      summary: Returns cluster configuration in HOCON format.
      description: Returns cluster configuration in HOCON format.
      operationId: getClusterConfiguration
      parameters: []
      responses:
        "200":
          description: Get cluster configuration
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: Incorrect configuration
        "500":
          description: Internal error
    put:
      tags:
      - clusterConfiguration
      summary: Updates cluster configuration in HOCON format.
      description: Updates cluster configuration in HOCON format.
      operationId: updateClusterConfiguration
      parameters: []
      requestBody:
        description: the cluster configuration to update.
        content:
          text/plain:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: Configuration updated
        "400":
          description: Incorrect configuration
        "500":
          description: Internal error
  /management/v1/configuration/cluster/{path}:
    get:
      tags:
      - clusterConfiguration
      summary: Returns configuration in HOCON format represented by path.
      description: Returns configuration in HOCON format represented by path.
      operationId: getClusterConfigurationByPath
      parameters:
      - name: path
        in: path
        description: to represent a cluster configuration.
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Configuration represented by path
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: Incorrect configuration
        "500":
          description: Internal error
  /management/v1/configuration/node:
    get:
      tags:
      - nodeConfiguration
      summary: Returns node configuration in HOCON format.
      description: Returns node configuration in HOCON format.
      operationId: getNodeConfiguration
      parameters: []
      responses:
        "200":
          description: Whole node configuration
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: Incorrect configuration
        "500":
          description: Internal error
    put:
      tags:
      - nodeConfiguration
      summary: Updates node configuration in HOCON format.
      description: Updates node configuration in HOCON format.
      operationId: updateNodeConfiguration
      parameters: []
      requestBody:
        description: the node configuration to update.
        content:
          text/plain:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: Configuration updated
        "400":
          description: Incorrect configuration
        "500":
          description: Internal error
  /management/v1/configuration/node/{path}:
    get:
      tags:
      - nodeConfiguration
      summary: Returns configuration in HOCON format represented by path.
      description: Returns configuration in HOCON format represented by path.
      operationId: getNodeConfigurationByPath
      parameters:
      - name: path
        in: path
        description: to represent a node configuration.
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Configuration represented by path
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: Incorrect configuration
        "500":
          description: Internal error